added blueprint to organize the file structure. emabled users to reset there pass word.
