+++ orange_it/__init__.py
@@ -16,7 +16,7 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
 db = SQLAlchemy(app)
 bcrypt = Bcrypt(app)
 login_manager = LoginManager(app)
-login_manager.login_view = 'login'
+login_manager.login_view = 'users.login'
 login_manager.login_message_category = 'info'
 app.config['MAIL_SERVER'] = 'smtp.gmail.com'
 app.config['MAIL_PORT'] = 587
@@ -26,4 +26,11 @@ app.config['MAIL_PASSWORD'] = dictionary_profile.email_profile.get('EMAIL_PASS')
 
 mail = Mail(app)
 
-from orange_it import routes
+from orange_it.users.routes import users
+from orange_it.posts.routes import posts
+from orange_it.main.routes import main
+
+app.register_blueprint(users)
+app.register_blueprint(posts)
+app.register_blueprint(main)
+
