+++ orange_it/main/routes.py
@@ -1,0 +1,17 @@
+
+from flask import render_template, request, Blueprint
+from orange_it.models import Post
+
+main = Blueprint('main', __name__)
+
+
+@main.route('/')
+def index():
+    page = request.args.get('page', 1, type=int)
+    posts = Post.query.order_by(Post.date_posted.desc()).paginate(page=page, per_page=5)
+    return render_template('index.html', posts=posts)
+
+
+@main.route('/about')
+def about():
+    return '<h1>About Page<h1'
