+++ orange_it/templates/auth/login.html
@@ -57,12 +57,12 @@
                                 {{form.submit(class="btn btn-outline-info")}}
                             </div>
                             <small class="test-muted ml-2">
-                                <a href="{{ url_for('reset_request') }}">Forgot Password?</a>
+                                <a href="{{ url_for('users.reset_request') }}">Forgot Password?</a>
                             </small>
                         </form>
                     </div>
              <small class="text-muted">
-                Do you need Account?<a class="ml-2" href="{{ url_for('register') }}">Sign In</a>
+                Do you need Account?<a class="ml-2" href="{{ url_for('users.register') }}">Sign In</a>
             </small>
             </div>
