+++ orange_it/templates/auth/reg.html
@@ -83,7 +83,7 @@
         </div>
         <div class="border-top pt-3">
             <small class="text-muted">
-                Already Have a Account?<a class="ml-2" href="{{ url_for('login') }}">Sign In</a>
+                Already Have a Account?<a class="ml-2" href="{{ url_for('users.login') }}">Sign In</a>
             </small>
         </div>
     </div>
