+++ orange_it/templates/base.html
@@ -26,12 +26,12 @@
         <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
             <div class="navbar-nav mi-auto">
                 {% if current_user.is_authenticated %}
-                    <a class="nav-item btn btn-primary" href="{{url_for('new_post')}}">New Post</a>
-                    <a class="nav-item btn btn-outline-primary mx-2" href="{{url_for('account')}}">Account</a>
-                    <a class="nav-item btn btn-outline-primary mx-2" href="{{ url_for('logout') }}">Logout</a>
+                    <a class="nav-item btn btn-primary" href="{{url_for('posts.new_post')}}">New Post</a>
+                    <a class="nav-item btn btn-outline-primary mx-2" href="{{url_for('users.account')}}">Account</a>
+                    <a class="nav-item btn btn-outline-primary mx-2" href="{{ url_for('users.logout') }}">Logout</a>
                 {% else %}
-                    <a class="nav-item btn btn-outline-primary mx-2" href="{{ url_for('login') }}">Login</a>
-                    <a class="nav-item btn btn-primary" href="{{url_for('register')}}">Sign Up</a>
+                    <a class="nav-item btn btn-outline-primary mx-2" href="{{ url_for('users.login') }}">Login</a>
+                    <a class="nav-item btn btn-primary" href="{{url_for('users.register')}}">Sign Up</a>
                 {% endif %}
             </div>
         </div>
