+++ orange_it/templates/user_posts.html
@@ -13,10 +13,10 @@
           <img class="rounded-circle article-img" src="{{ url_for('static', filename='profile_pics/' + post.author.image_file) }}">
           <div class="media-body">
             <div class="article-metadata">
-              <a class="mr-2" href="{{ url_for('user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
+              <a class="mr-2" href="{{ url_for('users.user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
               <small class="text-muted">{{ post.date_posted.strftime('%Y-%m-%d') }}</small>
             </div>
-            <h2><a class="article-title" href="{{ url_for('post', post_id=post.id) }}">{{ post.title }}</a></h2>
+            <h2><a class="article-title" href="{{ url_for('posts.post', post_id=post.id) }}">{{ post.title }}</a></h2>
             <p class="article-content">{{ post.content }}</p>
           </div>
         </article>
@@ -24,9 +24,9 @@
     {% for page_num in posts.iter_pages(left_edge=1,  right_edge=1, left_current=1, right_current=2) %}
         {% if page_num %}
             {% if posts.page==page_num %}
-                 <a class="btn btn-info mb-4" href="{{ url_for('user_posts', username = user.username, page=page_num) }}">{{ page_num }}</a>
+                 <a class="btn btn-info mb-4" href="{{ url_for('users.user_posts', username = user.username, page=page_num) }}">{{ page_num }}</a>
             {% else %}
-                 <a class="btn btn-outline-info mb-4" href="{{ url_for('index', username = user.username, page=page_num) }}">{{ page_num }}</a>
+                 <a class="btn btn-outline-info mb-4" href="{{ url_for('main.index', username = user.username, page=page_num) }}">{{ page_num }}</a>
             {% endif %}
         {% else %}
             ...
