+++ orange_it/users/utils.py
@@ -1,0 +1,35 @@
+
+import os
+import secrets
+from PIL import Image
+from flask import url_for
+from flask_mail import Message
+from orange_it import app, mail
+
+
+# todo delete old pic
+def save_picture(form_picture):
+    random_hex = secrets.token_hex(8)
+    _, f_ext = os.path.splitext(form_picture.filename)
+    picture_fn = random_hex + f_ext
+    picture_path = os.path.join(app.root_path, 'static/profile_pics', picture_fn)
+
+    output_size = (125, 125)
+    i = Image.open(form_picture)
+    i.thumbnail(output_size)
+    i.save(picture_path)
+
+    return picture_fn
+
+
+# todo connect with bulk mail
+def send_reset_email(user):
+    token = user.get_reset_token()
+    _user = user.username
+    msg = Message('Password Reset Request', sender='noreply@gmail.com', recipients=[user.email])
+    msg.body = f'''To reset your password, visit the following link:
+{url_for('users.reset_token', token=token, _external=True)}
+If you did not make this request than you can simply ignore this email and no change will be made. 
+'''
+    mail.send(msg)
+
